Combined the two main Open Data search functions to one
open_data_search and open_data_sentences are now combined in one function open_data_search with an additional parameter to turn output of detected sentences on and off.
As the main bottleneck in computation time is the detection of keywords (esp. the accession numbers), this step is now only computed once. This speeds things up by a factor of 2 compared to the previous version, when running both functions one after another.